requirements.txt
anyio==4.3.0
bidict==0.23.1
click==8.1.7
colorama==0.4.6
h11==0.14.0
httptools==0.6.1
idna==3.6
MetaTrader5==5.0.4200
numpy==1.26.4
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
PyYAML==6.0.1
simple-websocket==1.0.0
sniffio==1.3.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
